Added rescaling constant
Added a constant 1/50 for the scale of the map, and introduced it everywhere in the code that I saw it was needed.

Due to the rescaling, every constant value in the code that has physical units with a factor of (length)^n needs to be multiplied by (10*TerrainConstancts.MAP_SCALE)^n. In particular, anything with units of distance, velocity, or acceleration needs a factor of 10*TerrainConstancts.MAP_SCALE. The extra factor of 10 is to compensate for the old scale factor, which was previously being ignored.