Change to build with API compatibility against .NET 4.x
The previous setting was preferable, as it is more likely to be
supported on all platforms. However, Newtonsoft.Json fails
with 'PlatformNotSupportedException' related to reflection if compiled
without .NET 4.x. In the future we may have to try a Unity-specific fork
of Newtonsoft (https://github.com/jilleJr/Newtonsoft.Json-for-Unity) that
might solve the problem, but for now we just use a broader
API compatibility setting.