Simplify turret code by removing half-implemented state
Targets are now held only by the turret system. Turrets used
to have their own targeting state. This will make sense in the future
if we ever make them pick individual stuff independent of what
the unit as a whole is shooting. However, for now it was only
making the code more complicated without adding value.